more lazy refactor
