+++ src/akimbo/spark.py
@@ -1,15 +1,9 @@
-from typing import Callable, Iterable
-
 import awkward as ak
-import numpy as np
 import pyarrow as pa
 import pyspark
 from pyspark.sql.pandas.types import from_arrow_schema, to_arrow_schema
 
-from akimbo.apply_tree import run_with_transform
-from akimbo.mixin import LazyAccessor, match_any, numeric
-from akimbo.pandas import pd
-from akimbo.utils import to_ak_layout
+from akimbo.mixin import LazyAccessor
 
 sdf = pyspark.sql.DataFrame
 
@@ -28,7 +22,9 @@ class SparkAccessor(LazyAccessor):
         batches = data._collect_as_arrow()
         return pa.Table.from_batches(batches)
 
-    def to_output(self, data=None) -> pd.DataFrame | pd.Series:
+    def to_output(self, data=None):
+        from akimbo.pandas import pd
+
         # data is always arrow format internally
         data = self.to_arrow(data if data is not None else self._obj).to_pandas(
             types_mapper=pd.ArrowDtype
@@ -122,71 +118,6 @@ class SparkAccessor(LazyAccessor):
             return obj.mapInArrow(f, schema=from_arrow_schema(out_schema))
 
         return select
-
-    def __array_ufunc__(self, *args, where=None, out=None, **kwargs):
-        # includes operator overload like df.ak + 1
-        ufunc, call, inputs, *callargs = args
-        if out is not None or call != "__call__":
-            raise NotImplementedError
-
-        return self.__getattr__(ufunc)(*callargs, where=where, **kwargs)
-
-    def __getitem__(self, item) -> sdf:
-        def f(batches):
-            for batch in batches:
-                arr = ak.from_arrow(batch)
-                arr2 = arr.__getitem__(item)
-                if not arr2.fields:
-                    arr2 = ak.Array({"_ak_series_": arr2})
-                out = ak.to_arrow(
-                    arr2,
-                    extensionarray=False,
-                    list_to32=True,
-                    string_to32=True,
-                    bytestring_to32=True,
-                )
-                yield pa.RecordBatch.from_struct_array(out)
-
-        arrow_type = to_arrow_schema(self._obj.schema)
-        arr = pa.table([[]] * len(arrow_type), schema=arrow_type)
-        out1 = next(f([arr]))
-        out_schema = pa.table(out1).schema
-        return self._obj.mapInArrow(f, schema=from_arrow_schema(out_schema))
-
-    def transform(
-        self,
-        fn: callable,
-        *others,
-        where=None,
-        match=match_any,
-        inmode="array",
-        **kwargs,
-    ):
-        def f(arr, *others, **kwargs):
-            return run_with_transform(
-                arr, fn, match=match, others=others, inmode=inmode, **kwargs
-            )
-
-        return self.__getattr__(f)(*others, **kwargs)
-
-    def apply(self, fn: Callable, *others, where=None, **kwargs):
-        return self.__getattr__(fn)(*others, **kwargs)
-
-    @classmethod
-    def _create_op(cls, op):
-        def run(self, *args, **kwargs):
-            args = [
-                to_ak_layout(_) if isinstance(_, (str, int, float, np.number)) else _
-                for _ in args
-            ]
-            return self.transform(op, *args, match=numeric)
-
-        return run
-
-    def __dir__(self) -> Iterable[str]:
-        if self.subaccessor is not None:
-            return dir(self.subaccessors[self.subaccessor](self))
-        return super().__dir__()
 
 
 def concat_columns_zip_index(df1: sdf, df2: sdf) -> sdf:
